backend/requirements.txt
# FastAPI and Web Framework
fastapi
uvicorn
python-multipart
python-dotenv

# Database and Storage
supabase

# AI and ML
openai

# Data Processing
pydantic
pydantic-settings
httpx
requests

# Utilities
loguru